requirements_unpinned.txt
aiohappyeyeballs
aiohttp
aiosignal
anyio
attrs
certifi
charset-normalizer
colorama
coverage
datasets
dill
filelock
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
idna
iniconfig
Jinja2
joblib
MarkupSafe
mpmath
multidict
multiprocess
networkx
numpy
packaging
pandas
pluggy
propcache
pyarrow
Pygments
pytest
pytest-cov
python-dateutil
pytz
PyYAML
regex
requests
safetensors
scikit-learn
scipy
setuptools
six
sniffio
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3
xxhash
yarl

psycopg2-binary
